RF011#RF012#feat: update account transaction handling; implement AccountTransactionRequest, enhance deposit and withdrawal methods, and ensure balance validation

+++ src/main/java/com/Isabela01vSilva/bank_isabela/controller/response/account/AccountResponse.java
@@ -3,6 +3,7 @@ package com.Isabela01vSilva.bank_isabela.controller.response.account;
 import com.Isabela01vSilva.bank_isabela.domain.account.AccountStatus;
 import com.Isabela01vSilva.bank_isabela.domain.account.AccountType;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record AccountResponse(
@@ -10,7 +11,7 @@ public record AccountResponse(
         String numeroAgencia,
         AccountType tipoConta,
         AccountStatus statusConta,
-        Double saldo,
+        BigDecimal saldo,
         LocalDate dataCriacao
 ) {
 }
